// Fix bug in installer (ignore .svn folder when tmp img are deleted)
git-svn-id: http://dev.prestashop.com/svn/v1/branches/1.5.x@15403 b9a71923-0436-4b27-9f14-aed3839534dd

+++ classes/Tools.php
@@ -635,7 +635,7 @@ class ToolsCore
 		$dirname = rtrim($dirname, '/').'/';
 		$files = scandir($dirname);
 		foreach ($files as $file)
-			if ($file != '.' && $file != '..')
+			if ($file != '.' && $file != '..' && $file != '.svn')
 			{
 				if (is_dir($dirname.$file))
 					Tools::deleteDirectory($dirname.$file, true);
